Updated podspec for cocoapods index
We noticed that because there isn't a minimum sdk version set, the project is now breaking in iOS 4.3.

"QRadioItemElement.h:23:5: The current deployment target does not support automated __weak references"

Therefore I am updating the podspec in the index to have a minimum SDK target of 5.0.  Let me know if this is alright with you and I will pull request the change into cocoapods.