Fix QSegmentedElement display problem
Remove UIUserInterfaceIdiom specific code and instead make the
segmented control resize it's own width and height, with an initial
frame matching the cell.contentView.

Also add the segmentedcontrol to the cell's contentView rather than to
the cell directly.

This fixes the QSegmentedControl rendering issues that are visible in the iPhone and
iPad versions of the QuickDialog example (under the Sample Controls) section.