Fix editingChanged event not sent
Since iOS6 the UIControlEventEditingChanged is not triggered by programmatically changing the text property of UITextField. In my use case that is important , may be useful for the library. So we watch KVO notifications on the field. That is not very reliable as Apple may disable KVO at some time, but it works for iOS5~7